+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.3] - September 8th, 2020
+
+* Updated `TestStep` to be able to copy w/o the image.
+* Added `AsyncTestLoader` to provide more test loading options.
+* Updated with device info on the test reports.
+
+
 ## [1.0.2] - September 7th, 2020
 
 * Added `exit_app` step